[DO NOT MERGE] Re-write Storage
This is a rough-cut to sanity check the concept. Docstrings and full tests still missing.

New:
* `client.Storage` is an abstract base class.
* `locked_storage.LockedStorage` provides locking based on context managers and defaults to `threading.Lock`.

Changed:
* All `Storage` subclasses have been updated to use the new base class.
* `file.Storage` renamed to `file_storage.FileStorage`.
* `keyring_storage.Storage` renamed to `keyring_storage.KeyringStorage`.
* `django_orm.Storage` renamed to `django_orm.DjangoOrmStorage`.
* `KeyringStorage` and `FileStorage` are no longer threadsafe by default.

**Questionable changes**

As mentioned in [a comment on #344], `Credentials` has the unfortunate behavior of using the Storage's lock to prevent concurrently refreshing the credentials. IMO, Refresh locking should be handled as a separate concern. For now, `Storage` implements a no-op `lock()`.